Data.UUID.V4

Description

This module implements Version 4 UUIDs as specified in RFC 4122.

These UUIDs are generated from a psuedo-random generator. We use the System.Random StdGen as our random source.

All of the logic is encapsulated in the Random instance for the UUID type, so you are also free to use the random generator of your choice.

Documentation

nextRandom :: IO UUID

Generate a random UUID. Introduced in version 1.2.6.