Data.UUID.V5

Description

This module implements Version 5 UUIDs as specified in RFC 4122.

These UUIDs identify an object within a namespace, and are deterministic.

The namespace is identified by a UUID. Several sample namespaces are enclosed.

Documentation

generateNamed

Arguments

:: UUID
Namespace

-> [Word8]
Object

-> UUID

Generate a UUID within the specified namespace out of the given object.

Uses a SHA1 hash. The UUID is built from first 128 bits of the hash of the namespace UUID and the name (as a series of Word8).

namespaceDNS :: UUID

The namespace for DNS addresses

namespaceURL :: UUID

The namespace for URLs

namespaceOID :: UUID

The namespace for ISO OIDs

namespaceX500 :: UUID

The namespace for X.500 DNs